Updated SVG Icons.
In this edition a bunch of the icons got minor refinements, plus new versions and combinations were added.
I updated the naming of the graphics and removed the versioning. This way you can just swap the files and do not need to update each reference to the upgraded icons - just do a browser refresh.